Question Description
Objective
Today, we will learn about the Array data structure. Check out the Tutorial tab
for learning materials and an instructional video.

Task
Given an array, A, of N integers, print A's elements in reverse order as a single
line of space-separated numbers.

Example
A = [1,2, 3, 4|
Print Each integer is separated by one space.

Input Format
The first line contains an integer, N (the size of our array).
The second line contains N space-separated integers that describe array A's
elements.

Constraints
• 1< N< 1000
• 1 ≤ Ai≤ 10000, where Ali] is the ;th integer in the array.

Output Format
Print the elements of array A in reverse order as a single line of space
separated numbers.