Question Description
Objective
Today, we're delving into Inheritance. Check out the attached tutorial for learning materials and an instructional video.

Task
You are given two classes, Person and Student, where Person is the base class and Student is the derived class. Completed
code for Person and a declaration for Student are provided for you in the editor. Observe that Student inherits all the
properties of Person.

Complete the Student class by writing the following:
A Student class constructor, which has 4 parameters:
1. A string, first Name.
2. A string, last Name.
3. An integer, idNumber.
4. An integer array (or vector) of test scores, scores.

A char calculate() method that calculates a Student object's average and returns the grade character representative of their
calculated average: