Question Description
Task
Given an array, a, of size n distinct elements, sort the array in ascending order using the Bubble Sort algorithm above. Once
sorted, print the following 3 lines:

1. Array is sorted in numSwaps swaps.
where num Swaps is the number of swaps that took place.
2. First Element: firstElement
where first Element is the first element in the sorted array.
3. Last Element: lastElement
where last Element is the last element in the sorted array.
Hint: To complete this challenge, you will need to add a variable that keeps a running tally of all swaps that occur during
execution.

Example
a = [4, 3, 1, 2]
original a: 4 3 1 2
round 1 a: 3 1 2 4 swaps this round: 3
round 2 a . swaps this round: 2
round 3 a 1 2 3 4 swaps this round: 0

In the first round, the 4 is swapped at each of the 3 comparisons, ending in the last position. In the second round, the 3 is
swapped at 2 of the 3 comparisons. Finally, in the third round, no swaps are made so the iterations stop. The output is the
following:

Array is sorted in 5 swaps.
First Element: 1
Last Element: 4