Question Description
Task
Your local library needs your help! Given the expected and actual return dates for a library book, create a program that
calculates the fine (if any). The fee structure is as follows:
1. If the book is returned on or before the expected return date, no fine will be charged (i.e. fine = 0).
2. If the book is returned after the expected return day but still within the same calendar month and year as the expected
return date, fine = 15 Hackos × (the number of days late).
3. If the book is returned after the expected return month but still within the same calendar year as the expected return date,
the fine = 500 Hackos X (the number of months late).
4. If the book is returned after the calendar year in which it was expected, there is a fixed fine of 10000 Hackos.

Example
d1, ml, yl = 12312014 returned date
d2, m2, y2 = 112015 due date
The book is returned on time, SO no fine is applied.
d1, ml, yl = 112015 returned date
d2, m2, y2 = 12312014 due date
The book is returned in the following year, so the fine is a fixed 10000.

Input Format
The first line contains 3 space-separated integers denoting the respective day, month, and year on which the book was
actually returned.
The second line contains 3 space-separated integers denoting the respective day, month, and year on which the book was
expected to be returned (due date).

Constraints
• 1 ≤ D ≤ 31
• 1 < M < 12
• 1 ≤ Y ≤ 3000
It is guaranteed that the dates will be valid Gregorian calendar dates.

Output Format
Print a single integer denoting the library fine for the book received as input.